requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
daphne==3.0.2
Django==4.1.2
django-cors-headers==3.13.0
django-sse==0.4.1
djangorestframework==3.14.0
fonttools==4.39.2
future==0.18.3
hiredis==2.2.3
hupper==1.12
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==1.1.0
jsonfield==3.1.0
kiwisolver==1.4.4
kombu==5.3.1
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.4
pyudev==0.24.1
PyYAML==6.0
redis==4.5.3
seaborn==0.12.2
serial==0.0.97
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
sse==1.2
tabulate==0.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
zope.interface==6.0
